[2/2] Settings: On-The-Go mode
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

have fun!

Also:

[2/2] Settings: On-The-Go: Add quicktile

[2/2] Settings: On-The-Go: Hide on devices without camera
Hide camera switching for devices without front camera as well
if we dont have a front camera, open the first camera we have.
also hide the options from the dialog.

Conflicts:
	res/values/custom_strings.xml
	res/xml/power_menu_settings.xml

Conflicts:
	res/values/mahdi_strings.xml
	res/xml/power_menu_settings.xml
	src/com/android/settings/mahdi/PowerMenu.java
	src/com/android/settings/mahdi/quicksettings/QuickSettingsUtil.java